01-simple-agent/requirements.txt
fastapi==0.120.1
uvicorn[standard]==0.38.0
openai==2.6.1
python-dotenv==1.2.1
exa-py==1.16.1
mcp==1.19.0
anyio>=4.6
